Consolidate CLI test invocation
Shared CLI invocation keeps behavior tests focused on product assertions while preserving the same stdout/stderr capture behavior.

Constraint: Refactor work should stay behavior-preserving and reviewable as one PR with small commits.
Rejected: Keeping four local run_cli copies | it makes command harness changes noisy across unrelated tests.
Confidence: high
Scope-risk: narrow
Directive: Keep CLI invocation semantics centralized when adding command-level tests.
Tested: python3 -m unittest discover -s tests
Not-tested: N/A
